# src/main/resources/api/openapi.yml
openapi: 3.0.1
info:
  title: Purchase Approval System API
  description: Inbank Purchase Approval System API
  version: 1.0.0

paths:
  /api/v1/purchase/approval:
    post:
      summary: Decides whether to approve customer's request for a purchase
      operationId: purchaseApproval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerPurchaseRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerPurchaseResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Customer's personal ID not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    CustomerPurchaseRequest:
      type: object
      required:
        - personalId
        - amount
        - paymentPeriodInMonths
      properties:
        personalId:
          type: integer
          format: int64
          description: Customer's personal ID
          example: 12345678901
        amount:
          type: number
          minimum: 200.00
          maximum: 5000.00
          multipleOf: 0.01
          description: Customer's requested purchase amount
          example: 500.01
        paymentPeriodInMonths:
          type: integer
          description: Payment period for purchase
          example: 6
          minimum: 6
          maximum: 24
    CustomerPurchaseResponse:
      type: object
      properties:
        approved:
          type: boolean
          description: Purchase approval status
          example: true
        amount:
          type: number
          multipleOf: 0.01
          description: Amount of request purchase
          example: 500.01
        paymentPeriodInMonths:
          type: integer
          description: Payment period for purchase after approval decision
          example: 6
    ErrorResponse:
      type: object
      properties:
       code:
         type: string
         description: Server error code
         example: CUSTOMER_NOT_FOUND
       description:
         type: string
         description:  Detailed description of the error
         example: Lorum Ipsum